refactor: extract MinimalMessage into override system with hero avatar support
- Extract original MinimalMessage rendering into MinimalMessageDefault
- Make MinimalMessage a thin orchestrator using useComponentOverride
- Add fullAvatarUrl to message override props and BubbleMessage
- Update default template with proper CSS classes and avatar fallback
